test(cloudflare/vectorize): move worker fixture into fixtures/
