+++ packages/alchemy/test/Cloudflare/Vectorize/fixtures/vectorize-worker.ts
